requirements/base.txt
Django==5.0.7
kavenegar==1.1.2
python-dotenv==1.0.1